feat: move logger out into its own file, small refactors
