+++ src/server/index.ts
@@ -2,8 +2,10 @@ import { default as express, Express } from 'express';
 import { logger } from '@/server/lib/serverSideLogger';
 import { getConfiguration } from '@/server/lib/getConfiguration';
 import { applyMiddleware } from '@/server/lib/middleware';
+import redisClient from '@/server/lib/redis/redisClient';
+import { startBucketCapacityLogger } from '@/server/lib/rate-limit';
 
-const { port } = getConfiguration();
+const { port, rateLimiter, stage } = getConfiguration();
 
 const server: Express = express();
 
@@ -14,5 +16,13 @@ const serverInstance = server.listen(port);
 server.set('trust proxy', true);
 
 logger.info(`server running on port ${port}`);
+
+if (
+  stage !== 'DEV' &&
+  rateLimiter.enabled &&
+  typeof redisClient !== 'undefined'
+) {
+  startBucketCapacityLogger(redisClient);
+}
 
 export default serverInstance;
