+++ src/server/lib/rate-limit/index.ts
@@ -2,5 +2,8 @@ import rateLimit from './rateLimit';
 
 export type { RateLimiterConfiguration } from './types';
 
+export { startBucketCapacityLogger } from './logger';
+
 export * from './types';
+
 export default rateLimit;
